requirements.txt
ttkbootstrap
Pillow
requests